% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{supertrend-package}
\alias{supertrend}
\alias{supertrend-package}
\title{supertrend: SuperTrend Indicator and Signals for 'quantmod'}
\description{
Computes the SuperTrend technical indicator and extracts trading
signals from trend flips. Provides a 'quantmod'-idiomatic chart
overlay.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/your-account/supertrend}
  \item Report bugs at \url{https://github.com/your-account/supertrend/issues}
}

}
\author{
\strong{Maintainer}: Lars Bernhardsson \email{cran.sherry228@passinbox.com}

Authors:
\itemize{
  \item Lars Bernhardsson \email{cran.sherry228@passinbox.com}
}

}
\keyword{internal}
